Add introductory paragraph to intro issue template
Add an introductory paragraph to the "Introduction" issue template
(which is the template for the first issue to be assigned to a new
engineer on their onboarding board).

The aim of this introductory paragraph is to welcome the engineer and
introduce them to the onboarding process, how it works, and what they
can expect from it.

I deliberately haven't mentioned a timeline or how long onboarding might
take because I think that will vary from one engineer to the next.

Fixes https://github.com/hypothesis/onboarding/issues/33